pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "requests>=2", "cibuildwheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentsociety"
version = "1.3.0a0" # change it for each release
description = "LLM-based city environment agent building library"
authors = [
    { name = "Yuwei Yan", email = "pinkgranite86@gmail.com" },
    { name = "Junbo Yan", email = "yanjb20thu@gmail.com" },
    { name = "Jun Zhang", email = "zhangjun990222@gmail.com" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
]
requires-python = ">=3.9"
dependencies = [
    "aiodocker>=0.24.0",
    "asyncpg>=0.30.0",
    "click>=8.0.0",
    "dashscope>=1.14.1",
    "faiss-cpu>=1.9.0.post1",
    "fastapi>=0.103.1",
    "fastavro>=1.10.0",
    "geojson>=3.1.0",
    "grpcio>=1.67.1,<2.0.0",
    "json-with-comments>=1.2.10",
    "langchain-community>=0.3.13",
    "langchain-core>=0.3.28",
    "matplotlib>=3.8.3",
    "mlflow>=2.19.0",
    "mosstool>=1.3.0",
    "networkx>=3.2.1",
    "numpy>=1.20.0,<2.0.0",
    "openai>=1.58.1",
    "pandavro>=1.8.0",
    "protobuf<=4.24.0,<5.0.0",
    "psycopg[binary]>=3.2.3",
    "pycityproto>=2.2.7",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.2",
    "ray[default]>=2.40.0",
    "redis >=5.2.1",
    "requests>=2.32.3",
    "Shapely>=2.0.6",
    "sqlalchemy>=2.0.20",
    "torch>=2.5.1",
    "transformers>=4.47.1",
    "uvicorn>=0.23.2"
]

[project.urls]
Homepage = "https://github.com/tsinghua-fib-lab/agentsociety"
Repository = "https://github.com/tsinghua-fib-lab/agentsociety.git"
Issues = "https://github.com/tsinghua-fib-lab/agentsociety/issues"

[project.scripts]
agentsociety-sim = "agentsociety.cli.sim:agentsociety_sim"
agentsociety-ui = "agentsociety.cli.webui:cli"
# docker-runner = "agentsociety.cli.docker_runner:experiment_runner"
# runner = "agentsociety.cli.runner:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentsociety*"]

[tool.setuptools.package-data]
"agentsociety._dist" = ["*", "**/*"]

[tool.cibuildwheel]
build = ["cp3*manylinux_x86_64", "cp3*macosx_arm64"]
skip = ["cp36-*", "cp37-*", "cp38-*"]

[tool.cibuildwheel.linux]
repair-wheel-command = "auditwheel repair -w {dest_dir} {wheel}"

[tool.cibuildwheel.macos]
repair-wheel-command = "delocate-wheel -w {dest_dir} {wheel}"
